src/third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/dynamic-import/alpha/base-url.sub-expected.txt

This is a testharness.js-based test.
FAIL Relative URL-like from same origin classic <script> assert_equals: expected "gamma/import.js" but got "beta/import.js"
PASS Absolute URL-like from same origin classic <script>
FAIL Relative URL-like from cross origin classic <script> without crossorigin attribute assert_equals: Relative URL-like specifier resolution should fail expected "(unreached)" but got "beta/import.js"
PASS Absolute URL-like from cross origin classic <script> without crossorigin attribute
FAIL Relative URL-like from cross origin classic <script> with crossorigin attribute assert_equals: expected "gamma/import.js" but got "beta/import.js"
PASS Absolute URL-like from cross origin classic <script> with crossorigin attribute
PASS Relative URL-like from cross origin module <script>
PASS Absolute URL-like from cross origin module <script>
Harness: the test ran to completion.
